# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo API(Laravel 12)",
        "description": "This project is a simple RESTful API built with Laravel 12, focused on user authentication and checklist management. It provides features such as user registration and login, along with the ability to create and manage checklists and their items.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/checklist/{list_id}/item": {
            "post": {
                "tags": [
                    "Todo(ChecklistItem)"
                ],
                "summary": "Add new item to checklist",
                "description": "Create a new checklist item under the specified checklist. Requires Bearer token.",
                "operationId": "addChecklistItem",
                "parameters": [
                    {
                        "name": "list_id",
                        "in": "path",
                        "description": "Checklist ID to which the item will be added",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Buy milk"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Buy milk Description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Checklist"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Checklist not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/checklist/{list_id}/item/{id}": {
            "get": {
                "tags": [
                    "Todo(ChecklistItem)"
                ],
                "summary": "Get checklist item detail",
                "description": "Retrieve detailed information of a checklist item under the specified checklist. Requires Bearer token.",
                "operationId": "getChecklistItemDetail",
                "parameters": [
                    {
                        "name": "list_id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item detail retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChecklistItem"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Item or checklist not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Todo(ChecklistItem)"
                ],
                "summary": "Update checklist item",
                "description": "Update a specific item under a checklist. Requires Bearer token.",
                "operationId": "updateChecklistItem",
                "parameters": [
                    {
                        "name": "list_id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Update task name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Optional description",
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "ONGOING",
                                            "COMPLETED"
                                        ],
                                        "example": "COMPLETED"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Checklist item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ChecklistItem"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Checklist or item not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Todo(ChecklistItem)"
                ],
                "summary": "Delete a checklist item",
                "description": "Delete a specific item from a checklist. Requires Bearer token.",
                "operationId": "deleteChecklistItem",
                "parameters": [
                    {
                        "name": "list_id",
                        "in": "path",
                        "description": "Checklist ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Checklist Item ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist item deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Todo Item Deleted!"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Checklist or item not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/checklist/{id}/item": {
            "get": {
                "tags": [
                    "Todo(ChecklistItem)"
                ],
                "summary": "Get all checklist item",
                "description": "Retrieve detailed information of a checklist by its ID. Requires Bearer token.",
                "operationId": "getChecklistDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the checklist",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Checklist detail retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Checklist"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Checklist not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ChecklistItem": {
                "title": "Checklist Item",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Buy groceries"
                    },
                    "description": {
                        "type": "string",
                        "example": "Things to buy at the market"
                    },
                    "status": {
                        "type": "string",
                        "example": "ONGOING"
                    }
                },
                "type": "object"
            },
            "Checklist": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Daily Tasks"
                    },
                    "description": {
                        "type": "string",
                        "example": "Tasks to complete daily"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ChecklistItem"
                        }
                    },
                    "owner": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "properties": {
                    "username": {
                        "type": "string",
                        "example": "johndoe"
                    },
                    "email": {
                        "type": "string",
                        "example": "johndoe@example.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}